+++ packages/authenticator/amplify_authenticator/lib/src/controllers/authenticator_text_field_controller.dart
@@ -11,6 +11,5 @@ import 'package:flutter/widgets.dart';
 class AuthenticatorTextFieldController extends TextEditingController {
   AuthenticatorTextFieldController({super.text});
 
-  AuthenticatorTextFieldController.fromValue(TextEditingValue value)
-      : super.fromValue(value);
+  AuthenticatorTextFieldController.fromValue(super.value) : super.fromValue();
 }