+++ packages/authenticator/amplify_authenticator/lib/src/widgets/form_fields/confirm_sign_up_form_field.dart
@@ -138,6 +138,7 @@ class _ConfirmSignUpTextField extends ConfirmSignUpFormField<String> {
     this.controller,
   }) : super._();
 
+  @override
   final TextEditingController? controller;
 
   @override
@@ -220,6 +221,7 @@ class _ConfirmSignUpUsernameField
     this.controller,
   }) : super._();
 
+  @override
   final TextEditingController? controller;
 
   @override