+++ packages/authenticator/amplify_authenticator/lib/src/widgets/form_fields/verify_user_form_field.dart
@@ -74,6 +74,7 @@ class _VerifyUserTextField extends VerifyUserFormField<String> {
     this.controller,
   }) : super._();
 
+  @override
   final TextEditingController? controller;
 
   @override